Add health check route, remove custom boom
